fix static cast parce que axel veut faire le different

+++ Projet Runner 2D - Quentin - Lucas - Axel/HUD.cpp
@@ -19,13 +19,13 @@ void HUD::initTexture() {
 }
 
 void HUD::initHUD() {
-	goldPiece.setPosition({ 50, STGS::HEIGHT - 100 });
+	goldPiece.setPosition({ 50, static_cast<float>(STGS::HEIGHT - 100 )});
 	goldPiece.setSize({ 50,50 });
-	lifeholder.setPosition({ STGS::WIDTH / 2, 0 });
+	lifeholder.setPosition({ static_cast<float>(STGS::WIDTH / 2), 0 });
 	lifeholder.setSize({ 250,50 });
-	lvl.setPosition({ STGS::WIDTH/2,STGS::HEIGHT/2 });
+	lvl.setPosition({ static_cast<float>(STGS::WIDTH/2), static_cast<float>(STGS::HEIGHT/2) });
 	score.setPosition({ 0,0 });
-	timer.setPosition({ STGS::WIDTH-250, 0 });
+	timer.setPosition({ static_cast<float>(STGS::WIDTH - 250), 0 });
 	
 	score.setString("C'est le score");
 	lvl.setString("C'est le lvl");
